+++ draft-js-plugins-editor/src/Editor/index.js
@@ -2,13 +2,14 @@ import React, { Component } from 'react';
 import {
   Editor,
   EditorState,
+  DefaultDraftBlockRenderMap,
 } from 'draft-js';
 
 import createCompositeDecorator from './createCompositeDecorator';
 import moveSelectionToEnd from './moveSelectionToEnd';
 import proxies from './proxies';
 import * as defaultKeyBindingPlugin from './defaultKeyBindingPlugin';
-import { List } from 'immutable';
+import { List, Map } from 'immutable';
 
 /**
  * The main editor component
@@ -20,11 +21,13 @@ class PluginEditor extends Component {
     onChange: React.PropTypes.func.isRequired,
     plugins: React.PropTypes.array,
     defaultKeyBindings: React.PropTypes.bool,
+    defaultBlockRenderMap: React.PropTypes.bool,
     customStyleMap: React.PropTypes.object,
     decorators: React.PropTypes.array,
   };
 
   static defaultProps = {
+    defaultBlockRenderMap: true,
     defaultKeyBindings: true,
     customStyleMap: {},
     plugins: [],
@@ -225,6 +228,17 @@ class PluginEditor extends Component {
      ), {})
   );
 
+  resolveblockRenderMap = () => {
+    const blockRenderMap = this.props.plugins
+      .filter(plug => plug.blockRenderMap !== undefined)
+      .reduce((maps, plug) => maps.merge(plug.blockRenderMap), Map({}));
+    if (blockRenderMap && blockRenderMap.size > 0) {
+      return DefaultDraftBlockRenderMap.merge(blockRenderMap);
+    }
+
+    return DefaultDraftBlockRenderMap;
+  }
+
   resolveAccessibilityProps = () => {
     let accessibilityProps = {};
     const plugins = [this.props, ...this.resolvePlugins()];
@@ -259,6 +273,8 @@ class PluginEditor extends Component {
     const pluginHooks = this.createPluginHooks();
     const customStyleMap = this.resolveCustomStyleMap();
     const accessibilityProps = this.resolveAccessibilityProps();
+    const blockRenderMap = this.props.defaultBlockRenderMap ? this.resolveblockRenderMap() :
+    this.resolveblockRenderMap().merge(this.props.blockRenderMap);
     return (
       <Editor
         {...this.props}
@@ -266,6 +282,7 @@ class PluginEditor extends Component {
         {...pluginHooks}
         readOnly={this.props.readOnly || this.state.readOnly}
         customStyleMap={customStyleMap}
+        blockRenderMap={blockRenderMap}
         onChange={this.onChange}
         editorState={this.props.editorState}
         ref="editor"